add non-polyfilled entrypoint
